pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "postfix_sql_ucli"
authors = [
    {name = "Oleksandr Ostrenko", email = "oleksandr@ostrenko.dev"},
]
description = "uCLI administration tool for Postfix SQL backend"
readme = "README.rst"
requires-python = ">=3.8"
keywords = ["postfix", "sql", "administration"]
license = {text = "BSD-3-Clause"}
classifiers = [
    # As from http://pypi.python.org/pypi?%3Aaction=list_classifiers
    "Development Status :: 1 - Planning",
    # "Development Status :: 2 - Pre-Alpha",
    # "Development Status :: 3 - Alpha",
    # "Development Status :: 4 - Beta",
    # "Development Status :: 5 - Production/Stable",
    # "Development Status :: 6 - Mature",
    # "Development Status :: 7 - Inactive",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Operating System :: POSIX",
    "Operating System :: MacOS",
    "Operating System :: Unix",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "click",
    "passlib>=1.7",
    "sqlalchemy",
    "pyyaml",
    'importlib-metadata; python_version<"3.10"',
]
dynamic = ["version"]

[tool.black]
line-length = 120
target-version = ['py310']
preview = true
skip-string-normalization = true

[tool.setuptools.dynamic]
version = {attr = "postfix_sql_ucli.__version__"}

[project.optional-dependencies]
pdf = ["ReportLab>=1.2", "RXP"]
rest = ["docutils>=0.3", "pack ==1.1, ==1.3"]

[project.scripts]
postfix-sql-ucli = "postfix_sql_ucli.cli:main"

[tool.bumpversion]
allow_dirty = false
commit = false
message = "Bump version: {current_version} → {new_version}"
commit_args = ""
tag = false
sign_tags = false
tag_name = "v{new_version}"
tag_message = "Bump version: {current_version} → {new_version}"
current_version = "1.0.0"
parse = "(?P<major>\\d+)\\.(?P<minor>\\d+)\\.(?P<patch>\\d+)"
serialize = [
    "{major}.{minor}.{patch}"
]
search = "{current_version}"
replace = "{new_version}"

[[tool.bumpversion.files]]
filename = "src/postfix_sql_ucli/__init__.py"
search = "__version__ = \"{current_version}\""
replace = "__version__ = \"{new_version}\""

[[tool.bumpversion.files]]
filename = "CHANGELOG.md"
search = "Unreleased"

[[tool.bumpversion.files]]
filename = "CHANGELOG.md"
search = "{current_version}...HEAD"
replace = "{current_version}...{new_version}"
